Create sqlmodel for agents and for users with wallet
