+++ apps/api/src/db/__init__.py
@@ -1,0 +1,3 @@
+from .setup import Session
+
+__all__ = ["Session"]
