+++ apps/api/src/server.py
@@ -6,7 +6,7 @@ from fastapi import APIRouter, FastAPI, Request
 from psycopg2.extensions import connection as Tconnection
 
 from . import logger
-from .db import (
+from .db.db import (
     create_runtime,
     get_runtime_for_agent,
     get_runtimes,
